refactor logic for v2 and v4 signature computation
