+++ lib/controllers/object.js
@@ -10,12 +10,12 @@ const S3Event = require('../models/event');
 const S3Object = require('../models/object');
 const { TaggingConfiguration } = require('../models/config');
 const {
-  aws4SignatureBodyParser,
   capitalizeHeader,
   once,
   utf8BodyParser,
   xmlBodyParser,
 } = require('../utils');
+const { aws4SignatureBodyParser } = require('../signature/v4');
 
 function triggerS3Event(ctx, eventData) {
   ctx.app.emit(
