+++ lib/middleware/authentication.js
@@ -1,16 +1,15 @@
 'use strict';
 
-const { createHmac, createHash } = require('crypto');
+const { createHmac } = require('crypto');
 const { mapKeys, pickBy } = require('lodash');
 
 const AWSAccount = require('../models/account');
 const S3Error = require('../models/error');
 const { RESPONSE_HEADERS } = require('./response-header-override');
-const {
-  encodeURIComponentRFC3986,
-  parseDate,
-  parseISO8601String,
-} = require('../utils');
+
+const v2 = require('../signature/v2');
+const v4 = require('../signature/v4');
+const { encodeURIComponentRFC3986, parseDate } = require('../utils');
 
 const SUBRESOURCES = {
   acl: 1,
@@ -108,15 +107,13 @@ module.exports = () =>
       .join('/');
 
     // begin parsing for each part of the signature algorithm and the rest of the canonical request
-    let parseResult;
-    if (mechanisms.header) {
-      parseResult = parseHeader(ctx.headers);
-    } else if (mechanisms.queryV2) {
-      parseResult = parseQueryV2(ctx.query, ctx.headers);
-    } else if (mechanisms.queryV4) {
-      parseResult = parseQueryV4(ctx.query, ctx.headers);
-    }
-    const { signature, signatureProvided } = parseResult;
+    const request = mechanisms.header
+      ? parseHeader(ctx.headers)
+      : mechanisms.queryV2
+      ? v2.parseQuery(ctx.query)
+      : mechanisms.queryV4
+      ? v4.parseQuery(ctx.query)
+      : undefined;
 
     const canonicalizedQueryString = Object.entries(ctx.query)
       .filter(([param]) => {
@@ -130,14 +127,14 @@ module.exports = () =>
           return false;
         }
         return (
-          signature.version !== 2 ||
+          request.signature.version !== 2 ||
           SUBRESOURCES[param] ||
           RESPONSE_HEADERS[param]
         );
       })
       .map(
         ([param, value]) =>
-          signature.version === 2
+          request.signature.version === 2
             ? [param, value].slice(0, value ? 2 : 1).join('=') // v2 signing doesn't encode values in the signature calculation
             : [param, value].map(encodeURIComponent).join('='), // v4 signing requires the = be present even when there's no value
       )
@@ -166,49 +163,46 @@ module.exports = () =>
       amzHeaders: canonicalizedAmzHeaders,
     };
 
-    if (signature.version === 2) {
-      // S3 signing uses expiration time as timestamp
-      canonicalRequest.timestamp = parseResult.expires;
-    } else if (signature.version === 4) {
-      canonicalRequest.timestamp = parseResult.timestamp;
+    switch (request.signature.version) {
+      case 2:
+        // S3 signing uses expiration time as timestamp
+        canonicalRequest.timestamp = request.expires;
+        break;
+      case 4:
+        canonicalRequest.timestamp = request.time;
+        break;
     }
 
-    let stringToSign;
-    const account = AWSAccount.registry.get(signature.accessKeyId);
+    const account = AWSAccount.registry.get(request.accessKeyId);
     if (!account) {
       throw new S3Error(
         'InvalidAccessKeyId',
         'The AWS Access Key Id you provided does not exist in our records.',
-        { AWSAccessKeyId: signature.accessKeyId },
+        { AWSAccessKeyId: request.accessKeyId },
       );
     }
 
-    if (signature.version === 2) {
-      stringToSign = getStringToSignV2(canonicalRequest);
-
-      const signingKey = account.accessKeys.get(signature.accessKeyId);
-      const calculatedSignature = calculateSignatureV2(
-        stringToSign,
-        signingKey,
-        signature.algorithm,
+    if (request.signature.version === 2) {
+      request.stringToSign = v2.getStringToSign(canonicalRequest);
+      request.signingKey = account.accessKeys.get(request.accessKeyId);
+    } else if (request.signature.version === 4) {
+      request.stringToSign = v4.getStringToSign(canonicalRequest, request);
+      request.signingKey = v4.getSigningKey(
+        account.accessKeys.get(request.accessKeyId),
+        request.credential.date,
+        request.credential.region,
+        request.credential.service,
       );
-      if (signatureProvided === calculatedSignature) {
-        ctx.state.account = account;
-      }
-    } else if (signature.version === 4) {
-      stringToSign = getStringToSignV4(canonicalRequest, signature);
+    }
+    const calculatedSignature = createHmac(
+      request.signature.algorithm,
+      request.signingKey,
+    )
+      .update(request.stringToSign, 'utf8')
+      .digest(request.signature.encoding);
 
-      const secretKey = account.accessKeys.get(signature.accessKeyId);
-      const calculatedSignature = calculateSignatureV4(
-        stringToSign,
-        secretKey,
-        signature.credential.date,
-        signature.credential.region,
-        signature.credential.service,
-      );
-      if (signatureProvided === calculatedSignature) {
-        ctx.state.account = account;
-      }
+    if (request.signatureProvided === calculatedSignature) {
+      ctx.state.account = account;
     }
 
     if (!ctx.state.account) {
@@ -220,9 +214,9 @@ module.exports = () =>
           'The request signature we calculated does not match the signature ' +
             'you provided. Check your key and signing method.',
           {
-            AWSAccessKeyId: signature.accessKeyId,
-            StringToSign: stringToSign,
-            StringToSignBytes: Buffer.from(stringToSign)
+            AWSAccessKeyId: request.accessKeyId,
+            StringToSign: request.stringToSign,
+            StringToSignBytes: Buffer.from(request.stringToSign)
               .toString('hex')
               .match(/../g)
               .join(' '),
@@ -235,23 +229,27 @@ module.exports = () =>
   };
 
 function parseHeader(headers) {
-  const signature = {
-    version: undefined,
-    algorithm: undefined,
+  const request = {
+    signature: undefined,
     accessKeyId: undefined,
+    time: headers['x-amz-date'] || headers.date,
   };
-  let signatureProvided;
-  const timestamp = headers['x-amz-date'] || headers.date;
 
-  const [algorithm, ...components] = headers.authorization.split(' ');
+  const [algorithm] = headers.authorization.split(' ');
   switch (algorithm.toUpperCase()) {
     case 'AWS':
-      signature.version = 2;
-      signature.algorithm = 'sha1';
+      request.signature = {
+        version: 2,
+        algorithm: 'sha1',
+        encoding: 'base64',
+      };
       break;
     case 'AWS4-HMAC-SHA256':
-      signature.version = 4;
-      signature.algorithm = 'sha256';
+      request.signature = {
+        version: 4,
+        algorithm: 'sha256',
+        encoding: 'hex',
+      };
       break;
     default:
       throw new S3Error('InvalidArgument', 'Unsupported Authorization Type', {
@@ -260,98 +258,17 @@ function parseHeader(headers) {
       });
   }
 
-  if (signature.version === 2) {
-    if (components.length !== 1) {
-      throw new S3Error(
-        'InvalidArgument',
-        "Authorization header is invalid -- one and only one ' ' (space) required",
-        {
-          ArgumentName: 'Authorization',
-          ArgumentValue: headers.authorization,
-        },
-      );
-    }
-
-    const match = /([^:]*):([^:]+)/.exec(components[0]);
-    if (!match) {
-      throw new S3Error(
-        'InvalidArgument',
-        'AWS authorization header is invalid.  Expected AwsAccessKeyId:signature',
-        {
-          ArgumentName: 'Authorization',
-          ArgumentValue: headers.authorization,
-        },
-      );
-    }
-    signature.accessKeyId = match[1];
-    signatureProvided = match[2];
-  } else if (signature.version === 4) {
-    if (!('x-amz-content-sha256' in headers)) {
-      throw new S3Error(
-        'InvalidRequest',
-        'Missing required header for this request: x-amz-content-sha256',
-      );
-    }
-    if (
-      !headers['x-amz-content-sha256'].match(
-        /^(UNSIGNED-PAYLOAD|STREAMING-AWS4-HMAC-SHA256-PAYLOAD|[0-9A-Fa-f]{64})$/,
-      )
-    ) {
-      throw new S3Error(
-        'InvalidArgument',
-        'x-amz-content-sha256 must be UNSIGNED-PAYLOAD, STREAMING-AWS4-HMAC-SHA256-PAYLOAD, or a valid sha256 value.',
-        {
-          ArgumentName: 'x-amz-content-sha256',
-          ArgumentValue: headers['x-amz-content-sha256'],
-        },
-      );
-    }
-
-    // skip payload verification
-
-    const componentMap = new Map(
-      components
-        .join('')
-        .split(',')
-        .map((component) => {
-          const [key, ...value] = component.split('=');
-          return [key, value.join('=')];
-        }),
-    );
-
-    if (componentMap.size !== 3) {
-      throw new S3Error(
-        'AuthorizationHeaderMalformed',
-        'The authorization header is malformed; the authorization header ' +
-          'requires three components: Credential, SignedHeaders, and ' +
-          'Signature.',
-      );
-    }
-
-    for (const componentName of ['Credential', 'SignedHeaders', 'Signature']) {
-      if (!componentMap.has(componentName)) {
-        throw new S3Error(
-          'AuthorizationHeaderMalformed',
-          `The authorization header is malformed; missing ${componentName}.`,
-        );
-      }
-    }
-
-    // skip verification of each authorization header component
-
-    const [accessKeyId, date, region, service, termination] = componentMap
-      .get('Credential')
-      .split('/');
-
-    signature.accessKeyId = accessKeyId;
-    signature.credential = { date, region, service, termination };
-    signature.signedHeaders = componentMap.get('SignedHeaders').split(';');
-
-    signatureProvided = componentMap.get('Signature');
+  switch (request.signature.version) {
+    case 2:
+      Object.assign(request, v2.parseHeader(headers));
+      break;
+    case 4:
+      Object.assign(request, v4.parseHeader(headers));
+      break;
   }
 
   const serverTime = new Date();
-  const requestTime = parseDate(timestamp);
+  const requestTime = parseDate(request.time);
   if (isNaN(requestTime)) {
     throw new S3Error(
       'AccessDenied',
@@ -365,234 +282,10 @@ function parseHeader(headers) {
       'RequestTimeTooSkewed',
       'The difference between the request time and the current time is too large.',
       {
-        RequestTime: timestamp,
+        RequestTime: request.time,
         ServerTime: serverTime.toISOString().replace(/\.\d+/, ''),
       },
     );
   }
-  return {
-    signature,
-    signatureProvided,
-    timestamp,
-  };
-}
-
-function parseQueryV2(query) {
-  // authentication param names are case-sensitive
-  if (!('Expires' in query) || !('AWSAccessKeyId' in query)) {
-    throw new S3Error(
-      'AccessDenied',
-      'Query-string authentication requires the Signature, Expires and ' +
-        'AWSAccessKeyId parameters',
-    );
-  }
-
-  const signature = {
-    version: 2,
-    algorithm: 'sha1',
-    accessKeyId: query.AWSAccessKeyId,
-  };
-  const signatureProvided = query.Signature;
-
-  const serverTime = new Date();
-  const expiresTime = new Date(Number(query.Expires) * 1000);
-  if (isNaN(expiresTime)) {
-    throw new S3Error(
-      'AccessDenied',
-      `Invalid date (should be seconds since epoch): ${query.Expires}`,
-    );
-  }
-
-  if (serverTime > expiresTime) {
-    throw new S3Error('AccessDenied', 'Request has expired', {
-      Expires: expiresTime.toISOString().replace(/\.\d+/, ''),
-      ServerTime: serverTime.toISOString().replace(/\.\d+/, ''),
-    });
-  }
-
-  return {
-    signature,
-    signatureProvided,
-    expires: Number(query.Expires),
-  };
-}
-
-function parseQueryV4(query) {
-  // query param values are case-sensitive
-  if (query['X-Amz-Algorithm'] !== 'AWS4-HMAC-SHA256') {
-    throw new S3Error(
-      'AuthorizationQueryParametersError',
-      'X-Amz-Algorithm only supports "AWS4-HMAC-SHA256"',
-    );
-  }
-
-  const signature = {
-    version: 4,
-    algorithm: 'sha256',
-    accessKeyId: undefined,
-  };
-
-  if (
-    !('X-Amz-Credential' in query) ||
-    !('X-Amz-Signature' in query) ||
-    !('X-Amz-Date' in query) ||
-    !('X-Amz-SignedHeaders' in query) ||
-    !('X-Amz-Expires' in query)
-  ) {
-    throw new S3Error(
-      'AuthorizationQueryParametersError',
-      'Query-string authentication version 4 requires the ' +
-        'X-Amz-Algorithm, X-Amz-Credential, X-Amz-Signature, X-Amz-Date, ' +
-        'X-Amz-SignedHeaders, and X-Amz-Expires parameters.',
-    );
-  }
-  const [accessKeyId, date, region, service, termination] = query[
-    'X-Amz-Credential'
-  ].split('/');
-
-  signature.accessKeyId = accessKeyId;
-  signature.credential = { date, region, service, termination };
-  signature.signedHeaders = query['X-Amz-SignedHeaders'].split(';');
-
-  const signatureProvided = query['X-Amz-Signature'];
-
-  const timestamp = query['X-Amz-Date'];
-
-  const requestTime = parseISO8601String(timestamp);
-  if (isNaN(requestTime)) {
-    throw new S3Error(
-      'AuthorizationQueryParametersError',
-      "X-Amz-Date must be in the ISO8601 Long Format \"yyyyMMdd'T'HHmmss'Z'\"",
-    );
-  }
-
-  const expires = Number(query['X-Amz-Expires']);
-  if (isNaN(expires))
-    if (expires < 0) {
-      throw new S3Error(
-        'AuthorizationQueryParametersError',
-        'X-Amz-Expires must be non-negative',
-      );
-    }
-  if (expires > 604800) {
-    throw new S3Error(
-      'AuthorizationQueryParametersError',
-      'X-Amz-Expires must be less than a week (in seconds); that is, the ' +
-        'given X-Amz-Expires must be less than 604800 seconds',
-    );
-  }
-
-  const serverTime = new Date();
-  // NOTE: S3 doesn't care about time skew for presigned requests
-  const expiresTime = new Date(Number(requestTime) + expires * 1000);
-
-  if (serverTime > expiresTime) {
-    throw new S3Error('AccessDenied', 'Request has expired', {
-      'X-Amz-Expires': query['X-Amz-Expires'],
-      Expires: expiresTime.toISOString().replace(/\.\d+/, ''),
-      ServerTime: serverTime.toISOString().replace(/\.\d+/, ''),
-    });
-  }
-
-  return {
-    signature,
-    signatureProvided,
-    timestamp,
-  };
-}
-
-/**
- * Generates a string to be signed for signature version 2.
- *
- * @param {*} canonicalRequest
- */
-function getStringToSignV2(canonicalRequest) {
-  return [
-    canonicalRequest.method,
-    canonicalRequest.contentMD5,
-    canonicalRequest.contentType,
-    canonicalRequest.timestamp,
-    ...canonicalRequest.amzHeaders,
-    canonicalRequest.querystring
-      ? `${canonicalRequest.uri}?${canonicalRequest.querystring}`
-      : canonicalRequest.uri,
-  ].join('\n');
-}
-
-/**
- * Generates a string to be signed for signature version 4.
- *
- * @param {*} canonicalRequest
- * @param {*} signature
- */
-function getStringToSignV4(canonicalRequest, { credential, signedHeaders }) {
-  const canonicalHeaders = signedHeaders
-    .map((header) => `${header}:${canonicalRequest.headers[header]}\n`)
-    .join('');
-
-  // skip computing content sha
-  // TODO: support staging, verifying, and committing content within store interface
-  const contentHash =
-    canonicalRequest.headers['x-amz-content-sha256'] || 'UNSIGNED-PAYLOAD';
-
-  const canonicalRequestString = [
-    canonicalRequest.method,
-    canonicalRequest.uri,
-    canonicalRequest.querystring,
-    canonicalHeaders,
-    signedHeaders.join(';'),
-    contentHash,
-  ].join('\n');
-
-  return [
-    'AWS4-HMAC-SHA256',
-    canonicalRequest.timestamp,
-    [
-      credential.date,
-      credential.region,
-      credential.service,
-      credential.termination,
-    ].join('/'),
-    createHash('sha256').update(canonicalRequestString).digest('hex'),
-  ].join('\n');
-}
-
-/**
- * Performs the calculation of an authentication code for a string using the specified key and
- * algorithm.
- *
- * @param {String} stringToSign the string representation of a canonical request
- * @param {String} signingKey a secret access key
- * @param {String} algorithm should be one of "sha1" or "sha256"
- */
-function calculateSignatureV2(stringToSign, signingKey, algorithm) {
-  const signature = createHmac(algorithm, signingKey);
-  signature.update(stringToSign, 'utf8');
-  return signature.digest('base64');
-}
-
-/**
- * Performs the calculation of an authentication code for a string using the specified key and
- * various components required by v4.
- *
- * @param {String} stringToSign the string representation of a canonical request
- * @param {String} secretKey a secret access key
- * @param {String} date received from the credential header
- * @param {String} region received From the credential header
- * @param {String} service received From the credential header
- */
-function calculateSignatureV4(stringToSign, secretKey, date, region, service) {
-  const dateKey = createHmac('sha256', 'AWS4' + secretKey)
-    .update(date)
-    .digest();
-
-  const regionKey = createHmac('sha256', dateKey).update(region).digest();
-
-  const serviceKey = createHmac('sha256', regionKey).update(service).digest();
-
-  const signingKey = createHmac('sha256', serviceKey)
-    .update('aws4_request')
-    .digest();
-
-  return createHmac('sha256', signingKey).update(stringToSign).digest('hex');
+  return request;
 }
